Move CustomIconPack methods and delete the class
